Build swappyVK so and static libraries and package them along swappy GL libraries.
Test: built using ./gradlew build and verified swappyVK include and libraries are in the final zip file.

+++ src/swappyVk/SwappyVk.cpp
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-#include "SwappyVk.h"
+#include <swappyVk/SwappyVk.h>
+
 #include <map>
 #include <unistd.h>
 
@@ -338,7 +339,7 @@ public:
         // TEMP CODE: LOG REFRESH DURATION AND RATE:
         double refreshRate = mRefreshDur;
         refreshRate = 1.0 / (refreshRate / 1000000000.0);
-        ALOGD("Returning refresh duration of %llu nsec (approx %f Hz)", mRefreshDur, refreshRate);
+        ALOGD("Returning refresh duration of %llu nsec (approx %f Hz)", (unsigned long long) mRefreshDur, refreshRate);
 
         *pRefreshDuration = mRefreshDur;
         return true;
@@ -761,7 +762,7 @@ public:
 
         double refreshRate = mRefreshDur;
         refreshRate = 1.0 / (refreshRate / 1000000000.0);
-        ALOGI("Returning refresh duration of %llu nsec (approx %f Hz)", mRefreshDur, refreshRate);
+        ALOGI("Returning refresh duration of %llu nsec (approx %f Hz)", (unsigned long long) mRefreshDur, refreshRate);
         return true;
     }
